+++ src/main/java/com/siapa/managedbean/lazymodel/AlimentoLazyModel.java
@@ -1,8 +1,3 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package com.siapa.managedbean.lazymodel;
 
 import com.siapa.managedbean.lazymodel.generic.GenericLazyModel;
@@ -13,23 +8,23 @@ import com.siapa.service.generic.GenericService;
  *
  * @author Joao
  */
-public class AlimentoLazyModel extends GenericLazyModel<Alimento, Long> {
+public class AlimentoLazyModel extends GenericLazyModel<Alimento, Integer> {
 
-    public AlimentoLazyModel(GenericService<Alimento, Long> service) {
+    public AlimentoLazyModel(GenericService<Alimento, Integer> service) {
         super(service);
     }
 
     @Override
     public Alimento getRowData(String rowKey) {
         for (Alimento alimento : getDatasource()) {
-            if(alimento.getIdAlimento().equals(Long.valueOf(rowKey)));
+            if (alimento.getIdAlimento().equals(Integer.valueOf(rowKey)));
             return alimento;
         }
         return null;
     }
 
     @Override
-    public Long getRowKey(Alimento element) {
+    public Integer getRowKey(Alimento element) {
         return element.getIdAlimento();
     }
